comparaison tube + socket

+++ socket_serveur.c
@@ -2,7 +2,28 @@
 
 #define NB_MAX 20
 
+
+void handler(int sig){
+    unlink("tube");
+    _exit(EXIT_SUCCESS);
+}
+
+
 int main(int argc, char const *argv[]) {
+
+    struct sigaction sa;
+    int rc;
+    memset(&sa, 0, sizeof(sa));
+    sa.sa_handler = handler;
+    sa.sa_flags = 0;
+    rc = sigaction(SIGINT, &sa, NULL);
+    if(rc < 0) {
+        perror("sigaction");
+        exit(1);
+    }
+
+    clock_t start,end;
+    double time_used;
     unlink("socket");
     struct sockaddr_un address;
     struct sockaddr_un addr_client;
@@ -28,16 +49,18 @@ int main(int argc, char const *argv[]) {
         perror("erreur listen");
         exit(4);
     }
-    int i=0;
-    while(i<10){
+    int lue=0;
+    while(1){
         size=sizeof(struct sockaddr_un);
         fd_accept=accept(fd_socket,(struct sockaddr*)&addr_client,&size);
-
-        size_t lu=read(fd_accept,buff,300);
-        printf("lu %ld\n",lu);
-        printf("buff %s\n",buff);
-        i++;
-		sleep(2);
+        start=clock();
+        while((lue=read(fd_accept,buff,300))!=0){
+            buff[lue]=0;
+        }
+        end=clock();
+        time_used=((double)(end-start))/CLOCKS_PER_SEC;
+        printf("Temps utilisé pour lire dans le tube : %f\n",time_used);
+        printf("J'ai lu %d\n",lue);
     }
     close(fd_accept);
     close(fd_socket);
